Change output and intermediate directories and fix some Level 4 warnings

+++ vis_milk2/constanttable.cpp
@@ -86,7 +86,7 @@ bool CConstantTable::GrabShaderData(ID3D11Device* pDevice)
         m_ConstantBuffers.push_back(constantBuffer);
     }
 
-    for (size_t i = 0; i < ShaderDesc.BoundResources; i++)
+    for (UINT i = 0; i < ShaderDesc.BoundResources; i++)
     {
         ShaderBinding shaderBinding;
         m_pReflection->GetResourceBindingDesc(i, &shaderBinding.Description);
@@ -96,9 +96,9 @@ bool CConstantTable::GrabShaderData(ID3D11Device* pDevice)
     return true;
 }
 
-int CConstantTable::GetVariablesCount()
+size_t CConstantTable::GetVariablesCount()
 {
-    int total = 0;
+    size_t total = 0;
     for (size_t i = 0; i < m_ConstantBuffers.size(); i++)
         total += m_ConstantBuffers[i].Variables.size();
 
@@ -202,7 +202,7 @@ bool CConstantTable::ApplyChanges(ID3D11DeviceContext* pContext)
     return false;
 }
 
-ShaderVariable* CConstantTable::GetVariableByIndex(UINT index)
+ShaderVariable* CConstantTable::GetVariableByIndex(size_t index)
 {
     for (size_t i = 0; i < m_ConstantBuffers.size(); i++)
     {
